+++ packages/plugin-packages/mashroom-http-proxy-add-access-token/src/AddAccessTokenHttpProxyInterceptor.ts
@@ -3,14 +3,14 @@ import type {IncomingMessageWithContext, MashroomLogger} from '@mashroom/mashroo
 import type {MashroomSecurityService, MashroomSecurityUser} from '@mashroom/mashroom-security/type-definitions';
 import type {MashroomHttpProxyInterceptor, HttpHeaders, QueryParams, MashroomHttpProxyRequestInterceptorResult, MashroomWsProxyRequestInterceptorResult} from '@mashroom/mashroom-http-proxy/type-definitions';
 
-export default class AddIdTokenHttpProxyInterceptor implements MashroomHttpProxyInterceptor {
+export default class AddAccessTokenHttpProxyInterceptor implements MashroomHttpProxyInterceptor {
 
-    constructor(private addBearer: boolean, private idTokenHeader: string, private targetUris: Array<string>) {
+    constructor(private addBearer: boolean, private accessTokenHeader: string, private targetUris: Array<string>) {
     }
 
     async interceptRequest(targetUri: string, existingHeaders: Readonly<HttpHeaders>, existingQueryParams: Readonly<QueryParams>,
                             clientRequest: Request, clientResponse: Response): Promise<MashroomHttpProxyRequestInterceptorResult | undefined | null> {
-        const logger = clientRequest.pluginContext.loggerFactory('mashroom.httpProxy.addIdToken');
+        const logger = clientRequest.pluginContext.loggerFactory('mashroom.httpProxy.addAccessToken');
         const securityService: MashroomSecurityService = clientRequest.pluginContext.services.security.service;
         const user = securityService.getUser(clientRequest);
         const addHeaders = this.addHeaders(targetUri, user, logger);
@@ -23,7 +23,7 @@ export default class AddIdTokenHttpProxyInterceptor implements MashroomHttpProxy
     }
 
     async interceptWsRequest(targetUri: string, existingHeaders: Readonly<HttpHeaders>, clientRequest: IncomingMessageWithContext): Promise<MashroomWsProxyRequestInterceptorResult | undefined | null> {
-        const logger = clientRequest.pluginContext.loggerFactory('mashroom.httpProxy.addIdToken');
+        const logger = clientRequest.pluginContext.loggerFactory('mashroom.httpProxy.addAccessToken');
         const securityService: MashroomSecurityService = clientRequest.pluginContext.services.security.service;
         const user = securityService.getUser(clientRequest as any);
         const addHeaders = this.addHeaders(targetUri, user, logger);
@@ -39,9 +39,9 @@ export default class AddIdTokenHttpProxyInterceptor implements MashroomHttpProxy
         if (!user) {
             return;
         }
-        const idToken = user.secrets?.idToken;
-        if (!idToken) {
-            logger.error('Not ID token found in user object');
+        const accessToken = user.secrets?.accessToken;
+        if (!accessToken) {
+            logger.error('Not access token found in user object');
             return;
         }
 
@@ -59,19 +59,19 @@ export default class AddIdTokenHttpProxyInterceptor implements MashroomHttpProxy
         }
 
         if (!match) {
-            logger.debug('Don\'t add ID token header to request because uri does not match', targetUri);
+            logger.debug('Don\'t add access token header to request because uri does not match', targetUri);
             return;
         }
 
-        logger.debug('Adding ID token header to request', targetUri);
+        logger.debug('Adding access token header to request', targetUri);
 
         if (this.addBearer) {
             return {
-                'Authorization': `Bearer ${idToken}`,
+                'Authorization': `Bearer ${accessToken}`,
             };
         } else {
             return {
-                [this.idTokenHeader]: idToken,
+                [this.accessTokenHeader]: accessToken,
             };
         }
     }