+++ packages/plugin-packages/mashroom-http-proxy-add-access-token/src/mashroom-bootstrap.ts
@@ -1,10 +1,10 @@
-import AddIdTokenHttpProxyInterceptor from './AddIdTokenHttpProxyInterceptor';
+import AddAccessTokenHttpProxyInterceptor from './AddAccessTokenHttpProxyInterceptor';
 
 import type {MashroomHttpProxyInterceptorPluginBootstrapFunction} from '@mashroom/mashroom-http-proxy/type-definitions';
 
 const bootstrap: MashroomHttpProxyInterceptorPluginBootstrapFunction = async (pluginName, pluginConfig) => {
-    const {addBearer, idTokenHeader, targetUris} = pluginConfig;
-    return new AddIdTokenHttpProxyInterceptor(addBearer, idTokenHeader, targetUris);
+    const {addBearer, accessTokenHeader, targetUris} = pluginConfig;
+    return new AddAccessTokenHttpProxyInterceptor(addBearer, accessTokenHeader, targetUris);
 };
 
 export default bootstrap;