+++ packages/plugin-packages/mashroom-http-proxy-add-access-token/test/services/AddAccessTokenHttpProxyInterceptor.test.ts
@@ -1,10 +1,10 @@
 
-import AddIdTokenHttpProxyInterceptor from '../../src/AddIdTokenHttpProxyInterceptor';
+import AddAccessTokenHttpProxyInterceptor from '../../src/AddAccessTokenHttpProxyInterceptor';
 
-describe('AddIdTokenHttpProxyInterceptor', () => {
+describe('AddAccessTokenHttpProxyInterceptor', () => {
 
     it('adds headers if the uri matches', async () => {
-        const interceptor = new AddIdTokenHttpProxyInterceptor(false, 'X-USER-ID-TOKEN', ['foo', 'ba.r']);
+        const interceptor = new AddAccessTokenHttpProxyInterceptor(false, 'X-USER-ACCESS-TOKEN', ['foo', 'ba.r']);
 
         const req: any = {
             pluginContext: {
@@ -14,7 +14,7 @@ describe('AddIdTokenHttpProxyInterceptor', () => {
                         service: {
                             getUser: () => ({
                                 secrets: {
-                                    idToken: 'XXXXXXXXXXXX'
+                                    accessToken: 'XXXXXXXXXXXX'
                                 }
                             }),
                         }
@@ -28,13 +28,13 @@ describe('AddIdTokenHttpProxyInterceptor', () => {
 
         expect(result).toEqual({
             addHeaders: {
-                'X-USER-ID-TOKEN': 'XXXXXXXXXXXX'
+                'X-USER-ACCESS-TOKEN': 'XXXXXXXXXXXX'
             }
         });
     });
 
     it('adds a bearer token if addBeaerer is true', async () => {
-        const interceptor = new AddIdTokenHttpProxyInterceptor(true, 'X-USER-ID-TOKEN', ['foo', 'ba.r']);
+        const interceptor = new AddAccessTokenHttpProxyInterceptor(true, 'X-USER-ACCESS-TOKEN', ['foo', 'ba.r']);
 
         const req: any = {
             pluginContext: {
@@ -44,7 +44,7 @@ describe('AddIdTokenHttpProxyInterceptor', () => {
                         service: {
                             getUser: () => ({
                                 secrets: {
-                                    idToken: 'XXXXXXXXXXXX'
+                                    accessToken: 'XXXXXXXXXXXX'
                                 }
                             }),
                         }
@@ -64,7 +64,7 @@ describe('AddIdTokenHttpProxyInterceptor', () => {
     });
 
     it('does not add headers if the uri doesn\' match', async () => {
-        const interceptor = new AddIdTokenHttpProxyInterceptor(false, 'X-USER-ID-TOKEN', ['foo', 'ba.r']);
+        const interceptor = new AddAccessTokenHttpProxyInterceptor(false, 'X-USER-ACCESS-TOKEN', ['foo', 'ba.r']);
 
         const req: any = {
             pluginContext: {
@@ -74,7 +74,7 @@ describe('AddIdTokenHttpProxyInterceptor', () => {
                         service: {
                             getUser: () => ({
                                 secrets: {
-                                    idToken: 'XXXXXXXXXXXX'
+                                    accessToken: 'XXXXXXXXXXXX'
                                 }
                             }),
                         }